+++ app/src/main/java/de/baumann/browser/activity/Settings_Menu.java
@@ -5,7 +5,6 @@ import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.Menu;
-
 import androidx.activity.EdgeToEdge;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -13,10 +12,13 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
+
 import java.lang.reflect.Type;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -31,6 +33,7 @@ public class Settings_Menu extends AppCompatActivity {
     private List<MenuItem> masterList;
     private AdapterSettingsMenu adapter;
     private SharedPreferences sharedPreferences;
+
     public static final String PREF_NAME = "MenuPreferences";
     public static final String KEY_LIST = "MenuList";
 
@@ -40,6 +43,7 @@ public class Settings_Menu extends AppCompatActivity {
         HelperUnit.initTheme(this);
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_settings_menu);
+
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(true);
@@ -49,65 +53,65 @@ public class Settings_Menu extends AppCompatActivity {
 
         RecyclerView recyclerView = findViewById(R.id.recyclerViewSettings);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
+
         adapter = new AdapterSettingsMenu(masterList);
         recyclerView.setAdapter(adapter);
 
         ItemTouchHelper.SimpleCallback simpleCallback = new ItemTouchHelper.SimpleCallback(
                 ItemTouchHelper.UP | ItemTouchHelper.DOWN, 0) {
+
             @Override
             public boolean onMove(@NonNull RecyclerView recyclerView,
                                   @NonNull RecyclerView.ViewHolder viewHolder,
                                   @NonNull RecyclerView.ViewHolder target) {
+                int fromPosition = viewHolder.getAdapterPosition();
+                int toPosition = target.getAdapterPosition();
 
-                // Hier getBindingAdapterPosition() nutzen statt getAdapterPosition()
-                int fromPosition = viewHolder.getBindingAdapterPosition();
-                int toPosition = target.getBindingAdapterPosition();
-                // Sicherstellen, dass beide Positionen gültig sind
-                if (fromPosition != RecyclerView.NO_POSITION && toPosition != RecyclerView.NO_POSITION) {
-                    adapter.onItemMove(fromPosition, toPosition);
-                    saveList(); // Speichert die neue Reihenfolge sofort
-                    return true;
-                }
-                return false;
+                // Nur die Liste im Speicher tauschen und visuell updaten
+                Collections.swap(masterList, fromPosition, toPosition);
+                adapter.notifyItemMoved(fromPosition, toPosition);
+                return true;
             }
-            @Override
-            public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {}
-        };
-        new ItemTouchHelper(simpleCallback).attachToRecyclerView(recyclerView);
 
+            @Override
+            public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
+                // Keine Aktion benötigt
+            }
+        };
+
+        ItemTouchHelper itemTouchHelper = new ItemTouchHelper(simpleCallback);
+        itemTouchHelper.attachToRecyclerView(recyclerView);
     }
-    private void saveList() {
-        SharedPreferences.Editor editor = sharedPreferences.edit();
-        editor.putString(KEY_LIST, new Gson().toJson(masterList));
-        editor.apply();
+
+    // Garantiert, dass die Liste gespeichert wird, wenn die Activity in den Hintergrund tritt oder geschlossen wird
+    @Override
+    protected void onPause() {
+        super.onPause();
+        saveList();
     }
-    public void loadList() {
+
+    private void loadList() {
+        Gson gson = new Gson();
         String json = sharedPreferences.getString(KEY_LIST, null);
         Type type = new TypeToken<ArrayList<MenuItem>>() {}.getType();
-        masterList = new Gson().fromJson(json, type);
+        masterList = gson.fromJson(json, type);
+
+        if (masterList == null) {
+            masterList = new ArrayList<>();
+        }
+    }
+
+    private void saveList() {
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        Gson gson = new Gson();
+        String json = gson.toJson(masterList);
+        editor.putString(KEY_LIST, json);
+        editor.apply(); // Nutzt asynchrones Schreiben im Hintergrund
     }
 
     @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_help, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(android.view.MenuItem menuItem) {
-        if (menuItem.getItemId() == android.R.id.home) {
-            finish();
-        }
-        if (menuItem.getItemId() == R.id.menu_help) {
-            Uri webpage = Uri.parse("https://codeberg.org/Gaukler_Faun/FOSS_Browser/wiki/Menu");
-            BrowserUnit.intentURL(this, webpage);
-        }
+    public boolean onSupportNavigateUp() {
+        getOnBackPressedDispatcher().onBackPressed();
         return true;
-    }
-
-    @Override
-    public void finish() {
-        saveList();
-        super.finish();
     }
 }