+++ app/src/main/java/de/baumann/browser/preferences/BasePreferenceFragment.java
@@ -59,9 +59,8 @@ public class BasePreferenceFragment extends PreferenceFragmentCompat {
 
         MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(requireContext());
         builder.setTitle(preference.getTitle());
-        builder.setIcon(R.drawable.icon_edit);
-        builder.setNegativeButton(R.string.app_cancel, null);
-        builder.setPositiveButton(R.string.app_ok, (dialog, i) -> {
+        builder.setPositiveButton(R.string.app_cancel, null);
+        builder.setNegativeButton(R.string.app_ok, (dialog, i) -> {
             String newValue = Objects.requireNonNull(input.getText()).toString();
             if (preference.callChangeListener(newValue)) {
                 preference.setText(newValue);